renamed also epic to mapexport
